+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.0.0
+
+- removes deprecated `getEntitlements`, replaced by `getOfferings`
+- removes deprecated `makePurchase`, replaced by `purchaseProduct`
+
+- iOS: 
+    - added new method, `syncPurchases`, that enables syncing the purchases in the local receipt with the backend without risking a password prompt. The method was already available on Android.
+    - added a new method, `presentCodeRedemptionSheet`, for offer codes redemption.
+- Bumped purchases-hybrid-common to 1.5.0 [Changelog here](https://github.com/RevenueCat/purchases-hybrid-common/releases/1.5.0)
+- Bumped purchases-ios to 3.9.2 [Changelog here](https://github.com/RevenueCat/purchases-ios/releases/3.9.1)
+- Bumped purchases-android to 4.0.1 [Changelog here](https://github.com/RevenueCat/purchases-hybrid-common/releases/4.0.1)
+
 ## 1.4.3
 
 - Explicitly added `Purchases` framework to podspec's linker flags so it gets correctly linked in the plugin pod.